+++ CHANGELOG.md
@@ -41,6 +41,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - updated Dutch translation
 - updated Norwegian (Bokmål) translation
 - updated Italian translation
+- updated Finnish translation
 - updated to PHPUnit v12 (Thorsten)
 - migrated codebase to use PHP 8.3 language features (Thorsten)
 - migrated from WYSIWYG editor from TinyMCE to Jodit Editor (Thorsten)